# Speed up Golang compile times.