Previously our Golang compile was quite slow. This PR serves to improve this by splitting the singular code output file into multiple by splitting it on pre-determined points.

Build time changes as measured on a 12 core Ryzen machine.

Windows
360 seconds -> 132 seconds

Linux
249 seconds -> 88 seconds

Mac
20 seconds -> 17 seconds